pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nodetool-core"
version = "0.6.3-rc.9"
description = "Nodetool Core is the core library for Nodetool, providing the necessary functionality for building and running AI workflows."
readme = "README.md"
authors = [{ name = "Matthias Georgi", email = "matti.georgi@gmail.com" }]
requires-python = ">=3.10.13"

dependencies = [

    # Core dependencies
    "pydantic>=2.11.2",
    "python-multipart>=0.0.17",
    "aiohttp>=3.13.1",  # 3.13.1 fixes zstd decompression for chunked streams
    "brotli>=1.2.0",  # Required by aiohttp 3.13+ for br encoding
    # aiodns (pulled in by aiohttp speedups) crashes at import-time with pycares 5.x:
    # AttributeError: module 'pycares' has no attribute 'ares_query_a_result'
    # Pin until upstream aiodns supports pycares 5.x.
    "pycares<5",
    "websockets>=14.1",
    "fastapi>=0.115.5",
    "fastmcp>=2.12.4",
    "openapi-pydantic",
    "uvicorn>=0.32.0",
    "msgpack>=1.1.0",
    "networkx>=3.4.2",
    "supabase>=2.15.0",
    "paramiko>=4.0.0",
    "psycopg[binary]>=3.2.9",
    "psycopg-pool>=3.2.6",
    "docker==7.1.0",

    # Utils dependencies
    "watchdog>=4.0.0",
    "aioconsole",
    "click>=8.1.7",
    "cryptography>=43.0.0",
    "feedparser",
    "ftfy>=6.3.1",
    "importlib_metadata",
    "importlib-resources",
    "joblib",
    "jsonpath-ng>=1.7.0",
    "jsonschema",
    "attrs>=23.2.0",
    "keyring>=25.5.0",
    "psutil>=6.1.0",
    "nvidia-ml-py>=13.580.65",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.0",
    "prompt_toolkit>=3.0.50",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "setuptools>=80.8.0",
    "tabulate>=0.9.0",
    "tomli>=2.2.1",
    "tomlkit",
    "tqdm>=4.67.0",


    # Document processing dependencies
    "bs4==0.0.2",
    "langchain-text-splitters==1.0.0",
    "html2text",
    "markitdown",
    "PyMuPDF>=1.25.2",
    "pymupdf4llm>=0.0.27",
    "jinja2==3.1.6",

    # AI dependencies
    "anthropic>=0.76.0",
    "claude-agent-sdk>=0.1.19",
    "huggingface_hub>=0.35.3",
    "openai>=2.15.0",
    "traceloop-sdk>=0.50.0",
    "fal-client>=0.8.1",
    "gguf>=0.17.1",
    "ollama>=0.6.1",
    "tiktoken",
    "sentencepiece",
    "google-genai>=1.57.0",
    # "replicate==1.0.3",
    "runpod",
    "mcp",

    # Media dependencies
    "ffmpeg-python>=0.2.0",
    "pillow < 11.0.0",                    # due to mflux
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "pydub>=0.25.1",
    "opencv-python-headless>=4.11.0.86",

    # Data dependencies
    "numpy>=2.0.0",
    "pandas>=2.2.3",
    "llama-index-core>=0.14.4",
    "llama-index-embeddings-ollama>=0.8.3",
    "matplotlib==3.10.8",
    "seaborn>=0.13.2",
    "pypandoc>=1.14.0",
    "python-docx==1.2.0",
    "pdfplumber>=0.11.4",
    "paddleocr>=3.2.0",
    "paddlepaddle",
    "pytesseract>=0.3.13",
    "openpyxl>=3.1.3",

    # Database and storage
    "chromadb>=1.4.0",
    "aiofiles>=24.1.0",
    "aiosqlite>=0.20.0",
    "boto3>=1.35.63",
    "pymemcache>=4.0.0",
    "protobuf",
    "moto>=5.1.4",

    # Web and networking
    "httpx>=0.27.2",
    "websocket-client>=1.8.0",
    "pyngrok>=7.2.8",
    "playwright>=1.40.0",
    "hf_xet>=0.0.1",
    "hf_transfer",

    # Authentication and security
    "annotated-types>=0.7.0",
    "authlib>=1.3.2",
    "itsdangerous",
    "PyJWT",

    # Observability
    "opentelemetry-distro==0.60b1",
    "opentelemetry-exporter-otlp==1.39.1",
]

[project.optional-dependencies]
dev = [
    "aioresponses>=0.7.8",
    "pytest-timeout>=2.3.1",
    "pytest-xdist>=3.6.0",   # Parallel test execution
    "mypy>=1.0.0",           # Type checking
    "ruff==0.14.13",           # Linting and formatting
    "ty>=0.0.7",             # Fast type checking
]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
nodetool = "nodetool.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/nodetool"]

# Ruff configuration for linting and formatting
artifacts = ["package_metadata/nodetool-core.json"]
[tool.ruff]
# Python version
target-version = "py311"
line-length = 120
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable useful rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # Ruff-specific rules
]

# Ignore specific rules (low-value or too noisy)
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top of file (sometimes needed)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "F401",   # imported but unused (common in __init__.py)
    "F811",   # redefined while unused
    "UP006",  # use built-in collection type annotations
    "UP035",  # deprecated typing aliases (Dict, List, etc.)
    "UP045",  # prefer `X | None` over `Optional[X]`
    "RUF002", # NBSP vs '-' nit in docstrings
    "RUF010", # explicit conversion flag in f-strings
    "SIM102", # combine nested ifs
    "SIM105", # prefer contextlib.suppress over try/except/pass
    "ARG002", # unused method/function argument
    "ARG001", # unused function argument (API handlers)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG", "S101", "F401"]  # Allow unused args, assert, and unused imports in tests
"**/__init__.py" = ["F401"]  # Allow unused imports in __init__
"**/conftest.py" = ["F401"]  # Allow unused imports in conftest

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Respect magic trailing comma
skip-magic-trailing-comma = false
# Automatically detect the appropriate line ending
line-ending = "auto"

# Mypy configuration for type checking
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true

# Per-module options
[[tool.mypy.overrides]]
module = [
    "tests.*",
]
disallow_untyped_defs = false
disallow_incomplete_defs = false

# Basedpyright configuration for type checking
[tool.basedpyright]
typeCheckingMode = "basic"
exclude = [
    "tests/**",
]
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownArgumentType = "none"
reportUnknownParameterType = "none"
reportUnknownLambdaType = "none"
reportMissingTypeStubs = "none"
reportUnusedImport = "none"
reportUnusedVariable = "none"
reportUnusedFunction = "none"
reportUnusedClass = "none"
reportShadowedImports = "none"

[dependency-groups]
dev = [
    "basedpyright>=1.37.1",
]
